pyproject.toml
[project]
name = "tts-radio"
version = "0.1.0"
description = "Personalized 1-minute AI radio briefing service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "edge-tts>=7.2,<8.0",
    "fastapi>=0.115,<1.0",
    "gTTS>=2.5,<3.0",
    "httpx>=0.27,<1.0",
    "pydub>=0.25,<1.0",
    "pydantic>=2.7,<3.0",
    "python-dotenv>=1.0,<2.0",
    "streamlit>=1.44,<2.0",
    "uvicorn[standard]>=0.30,<1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
